Move NavigationSection to own file
